requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
deprecation==2.1.0
fastapi==0.110.1
Flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
packaging==24.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-keycloak==3.12.0
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
